+++ NEWS.md
@@ -1,6 +1,9 @@
 # Version 0.0 - Release Notes
 
 * [0.0.7]
+    * replaced separate beddows, embrs1 and embrs2 vehicle models with one 
+    model with extra argument method = "beddows"; in embrs_ice, etc. 
+    [2023-06-05] 
     * replaced ef_eea2019 with ef_vein_eea [2023-06-02]
     * added verbose argument to all ef functions and build_inventory; 
     added method documentation to EF output [2023-05-26]